chore:  move unit tests to tests folder
