+++ packages/v-slick-carousel/tests/setup-tests.ts
