+++ packages/v-slick-carousel/vite.config.ts
@@ -37,6 +37,6 @@ export default defineConfig({
   test: {
     globals: true,
     environment: 'jsdom',
-    setupFiles: ['./lib/setup-tests.ts']
+    setupFiles: ['./tests/setup-tests.ts']
   }
 })
